fix: email could be silently dropped if allowPlusSign is set to false with no fallback email
in order to ensure all email is properly handled, either bounce back or forward to the specific
email address, email with alias (i.e. foo+bar@baz.com) has to be treated as if foo@baz.com which
should already have a matching email address. In the future, we could provide an optional
synchronised lambda function to check for email alias and bounce back.