ci: enable validate workflow only to repository owner
The validate workflow was desgined as a way to validate changes by deploying it on a test
environment and make manual test possible. Even though it is achievable by using the
pull_request_target event and make sure the environment could be deployed only with manual approval,
it undoubtedly still poses a huge risk as stated on
https://securitylab.github.com/research/github-actions-preventing-pwn-requests/. It is better to
rethink how it should be setup on both github and aws side before we enable it for pull request made
by non repository owner.