refactor(extensions): remove setup for options page and content script
Options page seems to be a less common setup on web extensions as it can be done with popup already.
While content script allows injecting custom element on input box, the first version of the maildog
extensions will only covers context menu and popup for generating or accessing email alias for
simplicity.